Finished v1 of the itourmobile seeder

+++ database/migrations/2018_03_21_183844_create_tour_stops_table.php
@@ -17,7 +17,7 @@ class CreateTourStopsTable extends Migration
             $table->increments('id');
 
             $table->unsignedInteger('tour_id');
-            $table->unsignedTinyInteger('order');
+            $table->unsignedSmallInteger('order');
 
             $table->string('title', 255);
             $table->text('description')->nullable();
